whisper_turbo_selfhosted/backend/requirements.txt
absl-py==2.2.0
accelerate==1.5.2
addict==2.4.0
aio-pika==9.5.5
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiormq==6.8.1
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.8.0
apscheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
backoff==2.2.1
bangla==0.0.2
black==25.1.0
blinker==1.9.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.37.18
botocore==1.37.18
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1

csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
cython==3.0.12
dateparser==1.1.8
decorator==5.2.1
dlinfo==2.0.0
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
espeakng-loader==0.2.4
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.18.0
flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
funcy==2.0
g2pkk==0.1.2
grpcio==1.71.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
inflect==7.5.0
isodate==0.7.2
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kokoro==0.9.2
langcodes==3.5.0
language-data==1.3.0
language-tags==1.2.0
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.43.0
loguru==0.7.3
marisa-trie==1.2.1
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
misaki==0.9.3
monotonic-alignment-search==0.1.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
onnxruntime==1.21.0
onnxruntime-gpu==1.21.0
openai-whisper==20240930
packaging==24.2
pamqp==3.3.0
pandas==1.5.3
pathspec==0.12.1
phonemizer-fork==3.3.2
pillow==11.0.0
piper-phonemize
piper-tts==1.2.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
propcache==0.3.0
protobuf==6.30.1
psutil==7.0.0
pyaudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
pypinyin==0.53.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyyaml==6.0.2
rdflib==7.1.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.23.1
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.11.4
segments==2.3.0
sentencepiece==0.2.0
setuptools==77.0.3
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.7.5
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.1
sudachidict-core==20250129
sudachipy==0.6.10
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
# torch==2.6.0+cu118
# torchaudio==2.6.0+cu118
# torchvision==0.21.0+cu118
tqdm==4.67.1
trainer==0.0.36
transformers==4.49.0
triton==3.2.0
typeguard==4.4.2
typer==0.15.2
typing-extensions==4.12.2
tzlocal==5.3.1
umap-learn==0.5.7
unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
wave==0.0.2
weasel==0.4.1
werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
